flush execution input by intializing it when datajob stack is created
