+++ datajob/datajob_stack.py
@@ -6,6 +6,7 @@ from aws_cdk.core import CfnOutput
 
 from datajob import logger
 from datajob.datajob_context import DataJobContext
+from datajob.datajob_execution_input import DataJobExecutionInput
 
 
 class DataJobStack(core.Stack):
@@ -43,6 +44,7 @@ class DataJobStack(core.Stack):
         self.include_folder = include_folder
         self.resources = []
         self.outputs = {}
+        self.execution_input = DataJobExecutionInput()
         self.context = None
 
     def __enter__(self):
